Explicitly set bash as shell
This should fix the file globbing error described in #1.
I tried to use GNU make's `wildcard` function but unfortunately it
didn't work because the expansion of the package file takes place before
it was generated, thus an empty argument was passed to `pacman -U`.
Using `.SECONDEXPANSION` didn't help and also made the Makefile hard to
read.

Closes #1.